requirements.txt
openai
pandas
google-generative-ai
google-api-core
python-dotenv
